Fix type narrowing

+++ engine/core/parser/DaeParser.cpp
@@ -81,7 +81,7 @@ void Survive::DaeParser::loadGeometry(std::ifstream &reader)
 			break;
 		} else if (line.find("input semantic") != -1)
 		{
-			int index = line.find("offset");
+			size_t index = line.find("offset");
 			if (index != -1)
 			{
 				char c = line[index + 8];
@@ -93,7 +93,7 @@ void Survive::DaeParser::loadGeometry(std::ifstream &reader)
 
 void Survive::DaeParser::parsePointsLine(std::string &line, std::vector<glm::vec3> &vertices)
 {
-	int index = line.find('>');
+	size_t index = line.find('>');
 	auto numbers = Util::split(line.substr(index + 1), ' ');
 
 	for (int i = 0; i < numbers.size(); i += 3)
@@ -107,7 +107,7 @@ void Survive::DaeParser::parsePointsLine(std::string &line, std::vector<glm::vec
 
 void Survive::DaeParser::parseTexturesLine(std::string &line, std::vector<glm::vec2> &textures)
 {
-	int index = line.find('>');
+	size_t index = line.find('>');
 	auto numbers = Util::split(line.substr(index + 1), ' ');
 
 	for (int i = 0; i < numbers.size(); i += 2)
@@ -120,7 +120,7 @@ void Survive::DaeParser::parseTexturesLine(std::string &line, std::vector<glm::v
 
 Survive::Model Survive::DaeParser::parseIndices(Loader &loader)
 {
-	int index = m_VertexData.indicesLine.find('>');
+	size_t index = m_VertexData.indicesLine.find('>');
 	auto numbers = Util::split(m_VertexData.indicesLine.substr(index + 1), ' ');
 
 	std::vector<float> resultPoints;
@@ -229,8 +229,8 @@ void Survive::DaeParser::normalizeWeights()
 
 std::vector<std::string> Survive::DaeParser::getData(std::string &line)
 {
-	int start = line.find('>');
-	int end = line.find_last_of('<');
+	size_t start = line.find('>');
+	size_t end = line.find_last_of('<');
 
 	if (end == -1)
 	{
@@ -292,12 +292,12 @@ Survive::DaeParser::getJoint(std::ifstream &reader, std::string &line, const std
 {
 	static const int OFFSET = 4;
 
-	const int startIndex = line.find("id=");
+	const size_t startIndex = line.find("id=");
 	std::string restOfLine = line.substr(startIndex + OFFSET);
-	const int length = restOfLine.find('\"');
+	const size_t length = restOfLine.find('\"');
 
 	std::string name = restOfLine.substr(0, length);
-	int index = std::find(jointNames.begin(), jointNames.end(), name) - jointNames.begin();
+	size_t index = std::find(jointNames.begin(), jointNames.end(), name) - jointNames.begin();
 
 	std::getline(reader, line);
