+++ engine/core/parser/DaeParser.h
@@ -38,7 +38,7 @@ namespace Survive
 	struct JointData
 	{
 		Joint rootJoint;
-		int numberOfJoints{};
+		size_t numberOfJoints{};
 	};
 
 	class DaeParser
